+++ test/components/column/column-grouped.puppeteer-spec.js
@@ -1,4 +1,4 @@
-const { getConfig, getComputedStyle } = require('../../helpers/e2e-utils.js');
+const getConfig = require('../../helpers/e2e-utils.js');
 
 describe('Column Grouped Puppeteer Tests', () => {
   describe('Column Grouped selection disable tests', () => {
@@ -38,7 +38,7 @@ describe('Column Grouped Puppeteer Tests', () => {
     });
   });
 
-  fdescribe('Ability to make a combined column chart and line chart tests', () => {
+  Sdescribe('Ability to make a combined column chart and line chart tests', () => {
     const url = 'http://localhost:4000/components/column-grouped/example-column-grouped-with-line-combined';
 
     beforeAll(async () => {
@@ -46,8 +46,7 @@ describe('Column Grouped Puppeteer Tests', () => {
     });
 
     it('should run visual test', async () => {
-      await page.waitForSelector('.dot');
-      //await page.waitForSelector('.automation-column-line-dot-2');
+      await page.waitForSelector('.dot');S
       const image = await page.screenshot(); 
       const config = getConfig('line-chart'); 
       expect(image).toMatchImageSnapshot(config);